move db functions into files
