+++ lib/LoginServer/index.js
@@ -20,7 +20,7 @@ const packet = require('./packet');
 const { NPSUserStatus } = require('./npsUserStatus');
 
 const { Connection } = require("../../src/Connection");
-const database = require("../database/db");
+const pool = require("../database");
 
 function npsGetCustomerIdByContextId(contextId) {
 
@@ -44,6 +44,15 @@ function npsGetCustomerIdByContextId(contextId) {
     return user.contextId === contextId
   })
   return userRecord[0]
+}
+
+async function updateSessionKey(customerId, sessionKey, contextId, connectionId) {
+  const sKey = sessionKey.substr(0, 16);
+  return pool.connect().then(pool.query(
+    `INSERT INTO sessions (customer_id, session_key, s_key, context_id, 
+    connection_id) VALUES ($1, $2, $3, $4, $5) ON CONFLICT (customer_id) DO UPDATE SET session_key = $2, s_key = $3, context_id = $4, connection_id = $5`,
+    [customerId, sessionKey, sKey, contextId, connectionId],
+  )).catch((e) => { console.trace(e); });
 }
 
 /**
@@ -71,7 +80,7 @@ async function userLogin(connection, data) {
   const customer = npsGetCustomerIdByContextId(userStatus.contextId);
 
   // Save sessionKey in database under customerId
-  await database.updateSessionKey(
+  await updateSessionKey(
     customer.customerId.readInt32BE(0),
     userStatus.sessionKey,
     userStatus.contextId,
