+++ src/MCServer.js
@@ -14,10 +14,10 @@
 // You should have received a copy of the GNU General Public License
 // along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
-const database = require('../lib/database/db');
 const { ConnectionMgr } = require('./connectionMgr');
 const { startTCPListener } = require('./listenerThread');
 const { logger } = require('./logger');
+const pool = require('../lib/database');
 
 const connectionMgr = new ConnectionMgr();
 
@@ -59,11 +59,26 @@ async function startServers() {
   logger.info('Listening sockets create successfully.');
 }
 
+/**
+ * Create the sessions database table if it does not exist
+ * @param {Function} callback
+ */
+async function createDB() {
+  return pool.connect().then(pool.query(
+    `CREATE TABLE IF NOT EXISTS sessions (customer_id INTEGER NOT NULL UNIQUE, 
+    session_key TEXT, s_key TEXT, context_id TEXT, connection_id INTEGER)`,
+    [],
+  )).catch((e) => {
+    logger.error(`Error connecting to database: ${e}`);
+    process.exit(1);
+  });
+}
+
 function run(configurationFile) {
   // Connect to database
   // Start the server listeners
   startServers(configurationFile)
-    .then(database.createDB)
+    .then(createDB)
     .catch((err) => {
       throw err;
     });
